Do not set moment locale in comment code
The locale is a global setting and should not be set in a component.
Instead, we should call `moment.locale(django.languageCode)` in one
central place (in the platform code).

The current version has the concrete issue that the locale is set to two
different values: First to `django.languageCode` and then to
`utils.translation.get_language()` on the first interaction.