refactor tiles
The flex implementation head the issue that the layout in incomplete
lines was broken:

    1 2 3 4
    5     6

    1 2 3 4
    5  6  7

We thought about using the bootstrap grid which does not have that
issue. That would have meant to move all layout into the template, which
we ultimately decided against (or at least postponed for now).

This new implementation uses `display: grid`. This is even newer than
flex: The [candiate recommendation](https://www.w3.org/TR/css-grid-1/)
has been published few weeks ago and the feature is not enabled in any
major browser (though [it soon will
be](http://caniuse.com/#search=grid)).

To test this, you should enable the experimental feature via a browser
flag.