Introduce external projects
External projects inherit from regular projects and thus contain a
blueprint, a module and a phase. To make the handling and styling more
robust the external projects duration is stored in its phase as start
and end date.

External projects are using a different dashboard form, which can be
accesed from the regular select blueprint view.

For convenience a `extproject_url` filter and a `is_external` tag have
been introduced for use in templates.