Make reports available for arbitrary entries
To make reporting available for arbitrary entries a new template tag
`react_reports` and an accompanying react function are introduced.
The template tag creates a link and the javascript to call the react
function to initialize the modal.
The modals is inserted as a new child of `body` to prevent problems when
the tag is used in f.ex. collapsible menus.