Enable signals for inheriting projects.
Currently the signals are registered explicitely for the Project base model. In
meinBerlin we inherit from the base model to enable external projects and
bplans. As Django signals do not take inheritance into account when handling
signals, it is required to register signals without binding to a sender and
check if the senders type is a subclass of Project manually.