Add a paging aware filterset.
By default FilterSets from django_filter are keeping every non filter
argument in the query string. This results a problem when paging is used
in addition to filters (see liqd/a4-meinberlin#403)
It is desired to reset the page to zero when applying a filter. This PR
introduces a PagedFilterSet which removes the page argument from the
query string.