use template filter to get translated blueprint name
We currently store `blueprint.title` as `project.typ` when creating a
project. Unfortunately, this will write the translation that is used for
the creator into the database.

This pull request adds a filter that maps a blueprint key to the
translated title. This way, we can store the key into the database and
get the ranslation in the template where it is displayed.