Import ugettext_lazy as _ if used as a class property
Class properties may be evaluated on module load time when the users
locale is not available yet. Thus it is necessary to use ugettext_lazy
to show a correct translation.

For consistency the ugettext_lazy method is imported as `_` if it is
required in a single case. This is in accordance with the django docs
[1,2].

ugettext_lazy can be used as a replacement of ugettext in most cases.
Except if string has to be changed (for example with `string.replace`)
within the Python code.
String operations on lazily translated strings in templates are not
affected and can be used as usual.

*Attention*: The changes of this commit may not be visible on local dev
systems. For unknown reasons i can't reproduce the failures occuring on
the -dev and -stage servers when using ugettext.

[1] https://docs.djangoproject.com/en/1.8/topics/i18n/translation
/#standard-translation
[2] https://docs.djangoproject.com/en/1.8/topics/i18n/translation/#lazy-
translation