Show a notice to login and connect if an email is already registered.
By default, if the email address is already registered with an existing
account, the user will be prompted to enter another email address, for
which a new account will be created. The new account will be linked to
the social account.
In meinBerlin, instead of showing the form to create the new account,
users will be notified that they should login to the existing account
and connect the social account from the account settings.

Attention: Providers that won't pass the users email address won't work
with this page, as in that case it would be mandatory for users to
provide an email address to be able to create a new, valid account.