Use predicates to prevent access to items if no phase has started yet
It is possible to create items before any phase has been started.
On the project_detail page a according message and no content is shown.
This is checked via the `PhaseDispatchMixin`.
As for items it is not possible to use the `PhaseDispatchMixin` a new
predicate has been introduces which tests if a phase has been started
yet. Be aware, that this does not imply an active phase.
If no phase has been started yet the items may not be accessed. But if
the particpations did finish the items should still be available.