Add overwritable get_languages method to emails
To allow for translated emails the overwriteable `get_languages` method
is introduced. It gets the receiver as an attribute an may be
overwritten to send Emails according to (not yet available) user
settings.
Currently the currently set language is used, which is set from the
browsers language settings if the call results from a user request or
from `settings.LANGUAGE_CODE`.