Fix email notifications
Moderators did not get any emails as there was a regression bug
introduced with the default email language.

This commit also enforces the use of the meinBerlin Action proxy when
using signals.

Note, that signals for Actions are not automatically bubbled down to
proxies, as discussed in https://code.djangoproject.com/ticket/9318