Fix ProjectContextDispatcher for CreateView
get_object is called on create when the ProjectContext is validated.
But get_object throws a AttributeError if is doesn't find a slug or pk.
This is just a temporary fix, as get_object might behave strange if an
url defines an url with a named 'pk' or 'slug' param