Pass module in addition to project to ProjectMixin / project detail views
The module that is rendered on the project detail page is passed to the
active phases view in addition to the project.
This is required as with multi modules it is not known which module of a
project is currently active.
The module may be None if no phase has been started yet.

This should be non breaking as the counterpart to the ProjectMixin is
always the PhaseDispatchMixin.