Adapt active_phase and days_left to multiple modules
As multiple modules may have phases that are active in the same time the
notion of the active phase of a project (and thus the days left) is no
longer distinct.
To allow for backwards compatibility the active phase has been redefined
to follow the last active phases logic.
Altough this changed the current logic this commit is non breaking, as
in existing projects with only a single module and non overlapping
phases backwards compatibility is kept.