Add ModuleDetail view that maay redirect to the ProjectDetail
The ModuleDetail view redirects to the project detail if the selected
module is currently displayed there.

Altough the ModuleDispatchMixin seems to be closer related to the
modules app it should be next to the PhaseDispatchMixin.
They share functionality and the ModuleDispatchMixin is using the same
logic (in reverse) to decide if a redirect to the project detail page
should happen. To make this dependenance obvious it was decided to move
it to the projects app.