Order phases by start_date
This change is due to a discussion with @xi where a consistent interface
was requested.
This commit is not conequent though, as it won't change the default
order for all phases calls. This would be breaking as it is currently
assumed that phases are ordered by weight (f.ex. in the individual
dashboards). I propose to change the default order for all phases in
another PR