Use configurable ProjectDashboard abstraction
The ProjectDashboard is the central entry point to get information from
a dashboard which is related to a certain project.
It supplies a get_progress, get_menu and get_components methods.
It differs from the components registry as it depends on a certain
project.
Furthermore it allows to overwrite the whole dashboard for bplan and
external projects.

Note: it may be discussed if get_menu should be part of the dashboard or
if it should stay in the mixin code while relying on the get_components
methods.