Add a typed project dashboard and overwrite create views
The create view for the bplan and external-project blueprint types are
explicetly overwritten to allow for the correct project type creation.
Furthermore it allows to provide the tile_image instead of the big
image.

The TypedProjectDashboard on the other hand completely replaces the
default dashboard and shows a single fancy component for external and
bplan projects.