Load unload_warning.js in base_dashboard_project.html
base_form* is only loaded for the FormComponentHelpers. So the
unload_warning.js should be moved to the parent
base_dashboard_project.html template.
I did now use extra_js with calls to block.super but maybe it would be
more robust if we'd always include the script tag from the
base_dashboard_project.html body?