only show feedback button on large screens
fixes #518

We already did hide the feedback button on small screens. But there was
a rather large area in which we could not be reasonably sure that
nothing is covered. This is fixed now.

Note that the feedback form is already linked in the footer, so the
information is still available.