Revert to npm install
Fixes #799

I am not relatlivly sure that `npm install` is what we want. I am not
sure whether it is bug-free, but that is another question.

package-lock.json should not do any harm, but I do not see any benefit
in committing it to the repo. We did not have trouble with different
versions on different systems so far and it might cause issues when
using `npm link`. So I added it to .gitignore.