Enable scrolling in long modals
The core problem was a missing `overflow-y: auto;` on the modal.

It was missing because the body was missing the `modal-open` class which
is set via the Bootstrap javascript when a modal is opened. As we did
not initialize the modal via javascript but rather tried to simply use
the classes directly on the html elements the scrolling has not been
enabled.

Using the javascript initialization also fixes the need for the custom
creation of a backdrop, as this will created and removed automatically
by bootstrap.

Furthermore the modal was not added to the body but rather somewhere in
the dom tree which is advised against by the bootstrap docs [1]

[1]:
> Modals use position: fixed, which can sometimes be a bit particular
> about its rendering. Whenever possible, place your modal HTML in a
> top-level position to avoid potential interference from other
> elements. You’ll likely run into issues when nesting a .modal within
> another fixed element.