Hide dropdown menu and disable button for external projects
Fixes #817

As the duplication code will be part of the a4 dashboard it would be
very complex to make the duplication code extensible and allow for
correct duplication of external and bplan projects.

Duplication of external and bplan projects is still possible but the
duplicated project looses its "type".

This PR simply hides the duplication from the user.