Make ProjectContext a mixin
Mixins should not depend on their order on the class hierarchy.

By using @properties instead of setting them directly in the dispatcher
we don't have to think about class hirearchy orderings.
This was made possible due to the changes to filters which dropped the
requirement to set request.module|project

The module|project validation was dropped since we have no case where we
set the module|project from the url with a hierarchical e.g. idea_slug
from which another module|project may result.