Delete statements from archived bplans
After a bplan is archived it's related statements may be deleted.
For simpler development/deployment the auto_archive script is extended
to delete the statements, altough it may have been possible to add
another command.

To prevent from loosing statements that are created just before the
participation ends, the deletion is delayed for 48 hours.
(In rare cases a statement may not have been sent due to the async
nature of our email interface, when the project is already archived)