Add container detail page
As the default project detail is not suitable to show containers there
have been 4 options considered:

1. Register a container-detail view with an url and link to it directly
from the tile.
2. Overwrite the a4/projects/urls.py from meinBerlin and route
"/projects/*/" to a custom view where the dispatcher is overwritten to
conditionally redirect.
3. Modifiy the project_detail template used in meinBerlin.
4. Add a dummy phase to each container (1970-2100) which is connected to
the container detail view.

We decided to use option 3 because it seemed the less invasive while it
kept the url structure.