Modify generated category form field instead of reinitialize it
The category fields had not been translated as the field had been
reinitialized instead of modified. With this PR the auto generated field
(with its localized verbose_name) will be kept and adapted to the
filtered queryset.

Furthermore the required parameter is set to true if there are any
categories for the module.