Fixt tests (clear caches)
python lru_caches may be invalid when using a single python instance in
which the database is reset between calls.

lru_cache uses `hash` to identify objects and django models are
returning the object.pk as the hash. As the database is reset for every
test function each projects pk is 1 and the cache is invalid. It always
returns the first created projects status.