Fix poll dashboard
The `objects.update_or_create` fails on postgres if a default manager is
used which returns an annotated queryset by default.
The error is `psycopg2.NotSupportedError: FOR UPDATE is not allowed with
GROUP BY clause`

Unfortunately it was not possible to set an RelatedManager that is used
when accesing choices via a backreference (e.g. 'question.choices').

The solution described at
https://docs.djangoproject.com/en/1.11/topics/db/managers/#using-managers-for-related-object-access
did not work for me.

This fixes the error by avoiding setting the annotated queryset by
default. On the downside this means additional database calls for each
choice that is serialized.

If we will hit a performance problem, we will have to reinvestigate the
issue.