alias shariff in webpack config
Without this, the call to `require('shariff')` in app.js will pull in
the es6 sources of shariff. We tell babel to ignore most stuff from
node_modules (due to issues in the past), so it will not be transpiled.

The fix is to pull in the pre-build file instead.