Fix webpack
As shariff.complete.js does not export a valid module we would either
have to build shariff from the source, or use the global
`window.Shariff` variable.

We decided to use the global variable, as it poses a simple
understandable solution and does not require additional dependencies to
build the project.