Refactor SimpleItemExport and ItemExportView
The Idea is to have a single View function in the class hierarchy which
is coming from Abstract*ExportView.

The AbstractExportView requires the following additional methods:
- `get_base_filename()`
- `get_header()`
- `export_rows()`

The `AbstractExportView` may be used together with `BaseExport` which
adds support for `VirtualFieldMixin`s and provides the `get_header()`
and `export_rows()` methods.
It depends on the overwritable `get_fields()` method which should return
a tuple of a list of field names and a list of localized header string.
By default `get_fields()` returns all the fields defined by
`VirtualFieldMixin`s by calling the `get_field_data()` method which
tries to get the data from `get_FOO_data(object)` or `object.FOO`

The `BaseExport` class is not always used together with
AbstractXlsxExportView as may be seen in meinberlin in
`documents/exports.py`.
It may be discussed if the `get_field_data()` method should fall back to
`object.FOO` as this is actually behaviour from the
`MultipleObjectExport` class.

The `MultipleObjectExport` extends `BaseExport` and allows to set the
following properties analogous to e.g. Django forms.
- `fields`
- `exclude`
- `model`

It is based on the generic `MultipleObjectMixin` and thus allows to
overwrite `get_queryset()` to determine the to be exported objects.
Furthermore it always sets a Link to the object via `get_absolute_url()`
as an identifier as the first exported column. As `get_link_data()` is
using `self.request` this Method has to be used together with a View
class.

Finally the `BaseItemExportView` is connecting the AbstractView with the
`MultipleObjectMixin` and provides a `get_base_filename()`. It still
requires to be used together with `VirtualMixins` for the concrete
exports. There is a default implementation for `get_description_data`
which may be used if the concrete export is setting the `fields` to
contain `description`.
But i think this class should be moved back to meinBerlin again, as it
is very specific.

----

There is now also a `UserGeneratedContentExportMixin` which exports the
created date and the creators username. But we can not be sure if
username is set but we don't  want to use creator.get_username() as this
would leak the email addresses on meinberlin. I'd propose to move this
to meinBerlin, too