ratings: fix crash on double creation
When a user has two tabs with the same page open and rates an item
in both of them, this will crash in the second tab if the user
has not previously rated the item before page load.

The reason is that we decide based on information from page load
whether we call 'create' or 'update' of the rating. If we call
'create' twice, it crashes.

This patch fixes the issue by making the API check whether a
rating exists already and returning the corresponding ID if
it is the case. The frontend again uses the ID to issue
an 'update' request.

This requires an additional roundtrip to the user, but keeps
the internal logic as simple as possible.