modules: add properties for end, start and status
I solved it a bit differently from the stuff in the projects. The
projects rely on the phases being started/running/whatever. This will
become confusing once we have more than one phase in a module and there
is time between the phases. A module is also running between its phases.