rules: define view_item independently of phase start
The apps with phases that needed content added by a moderator prior to
the participation had their view_item rules defined depending on the
context start. As the rules were not used to determine if the phase
content was shown, it makes more sense to define them analogous to all
other view_item rules.