apps/users: Update registration form code
Use simple overwrites of the allauth forms, like we do on a+. Also,
activate the social registration form for the servicekonto, to mirror
what we do on a+ for github, twitter etc. This potentially fixes some
outstanding bugs and lowers the risk of self-introduced bugs. Follow
up commits might be needed after some testing.