apps/bplan: make office worker update email sent when bplan saved
The update_fields are not automatically known for the save signal. We
have to explicitly set them, which we do here
https://github.com/liqd/a4-meinberlin/blob/e2e1c21d7943c551b02504734271ea5f6dacef13/meinberlin/apps/bplan/management/commands/bplan_get_location.py#L26
and here
https://github.com/liqd/a4-meinberlin/blob/0441b664467f5c428013bd230c1f286f506409dd/meinberlin/apps/bplan/tasks.py#L78.
These cases atomatically update the bplan point and district when
something is changed and the if (update_fields... checks if the update
is actually automatically done by these methods or manually (via API or
dashboard). So, to send the email as result of a manual update, we need
the same if-statement.