webpack: Remove vendor chunk
Using the vendor chunk is, IIUC, a leftover from Webpack <= 3 times.
Webpack >= 4 promote a different kind of chunk splitting. This is
a step into that direction.

The probably biggest benefit is that there's now a single big `adhocracy4`
entry point which includes all base dependencies, avoid odd ordering issues
and making it clearer where to add stuff to.