apps/dashboard: filter dashboard project, bplan, extproject, container lists for group members - fixes #2616
It would have been much better to filter by permissions instead of doing
this dance, but as explained in issue #2616 it's not that easy. With
this we have to take care, when we add or remove project update
permissions, as this (instead of working with permissions) works with
knowing the roles and showing respective projects only to people allowed
to update them.