apps/polls: make sure detail view knows which view to display - fixes #743
depending on the timeline and the module clusters, either the project view with or without initial timeline tile or the module view need to be displayed.
This is needed  to work with the timeline/cluster logic made use of in mB.
https://github.com/liqd/a4-meinberlin/commit/a3e568f77c68dfb78845968a044545b212a5ae36
It should not break projects without that logic (opin) as there the modules are just not in a cluster.
The detail template is overwritten in opin.