templates: blocktrans blocks refactored in all apps where these were not in 1 line
Translations in po-files from blocktrans blocks contain linebreaks with \n in them if they are not in 1 line in the templates.
In order to reinstate a necessary blocktranslate in the cms//storefronts template, the html_date tag was removed, as it is not possible to use this in a blocktrans.