replace deprecated .eslintrc with eslint.config.mjs
I tried to include all the rules and plugins from the old config as
far as they are compatible with eslint 9.

eslint-config-standard and eslint-config-standard-react seem to be
dead and have been forked into neoconfig.The promise and n plugin is
included in neoconfig afaik and doesn't need to be added explicitly.